Module task_queue::spawn_policy

Trait that controls the number of threads.

The queue is trying to change the number of threads only when new tasks come into it. In the TaskQueue::enqueue method.

Structs

DynamicSpawnPolicy
Policy that provide dynamic number of threads for queue.

ManualSpawnPolicy
Policy that makes it possible control of the threads number manually

ManualSpawnPolicyController
Controller which makes possible control of the threads number

StaticSpawnPolicy
Policy that provide max number of threads for queue.

Traits

SpawnPolicy